# OpenAPISpecification.yaml
swagger: '2.0'
info:
  description: ' REST APIs for Gowma Online Store'
  version: '1.0'
  title: GOWMA service API
  termsOfService: Terms of service
  contact:
    name: Gagan Deep
    url: 'https://org.mihy.gowma/about/'
    email: gagandeepts@gmail.com
  license:
    name: Apache License Version 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0'
host: 'localhost:8081'
basePath: /gowma-service
...
paths:
  /api/v1/offers:
    get:
      tags:
        - offer-controller
      summary: Get a list of offers for category and product id
      operationId: getOffersForCategoryIdNProductIdUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: offer-category-id
          in: query
          description: offer-category-id
          required: true
          type: integer
          format: int32
        - name: offer-product-id
          in: query
          description: offer-product-id
          required: false
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Offer'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
        - offer-controller
      summary: Create a offer
      operationId: createProductUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: offer
          description: offer
          required: true
          schema:
            $ref: '#/definitions/Offer'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/Offer'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/api/v1/offers/{offer-id}':
    put:
      tags:
        - offer-controller
      summary: Update a offer by id
      operationId: updateProductUsingPUT
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: offer-id
          in: path
          description: offer-id
          required: true
          type: integer
          format: int32
        - in: body
          name: offer
          description: offer
          required: true
          schema:
            $ref: '#/definitions/Offer'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Offer'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - offer-controller
      summary: Delete a offer by id
      operationId: deleteProductUsingDELETE
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: offer-id
          in: path
          description: offer-id
          required: true
          type: integer
          format: int32
      responses:
        '204':
          description: No Content
          schema:
            $ref: '#/definitions/User'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /api/v1/products:
    post:
      tags:
        - product-controller
      summary: Create a product
      operationId: createProductUsingPOST_1
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: product
          description: product
          required: true
          schema:
            $ref: '#/definitions/Product'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/Product'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /api/v1/products/_search:
    post:
      tags:
        - product-controller
      summary: Performs search on products
      operationId: searchProductsUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: productSearchRequest
          description: productSearchRequest
          required: true
          schema:
            $ref: '#/definitions/ProductSearchRequest'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Product'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/api/v1/products/{product-id}':
    get:
      tags:
        - product-controller
      summary: Get a product by id
      operationId: getProductByIdUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: product-id
          in: path
          description: product-id
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Product'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    put:
      tags:
        - product-controller
      summary: Update a product with given id
      operationId: updateProductUsingPUT_1
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: product-id
          in: path
          description: product-id
          required: true
          type: integer
          format: int32
        - in: body
          name: product
          description: product
          required: true
          schema:
            $ref: '#/definitions/Product'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Product'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - product-controller
      summary: Delete a product with id
      operationId: deleteProductUsingDELETE_1
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: product-id
          in: path
          description: product-id
          required: true
          type: integer
          format: int32
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  '/api/v1/products/{product-id}/product-images':
    get:
      tags:
        - product-controller
      summary: Get a list of product images for a product with id
      operationId: getAllProductImagesByProductIdUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: product-id
          in: path
          description: product-id
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/ProductImage'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
        - product-controller
      summary: Create a list of Product Images for a product with id
      operationId: createProductImagesForProductIdUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: product-id
          in: path
          description: product-id
          required: true
          type: integer
          format: int32
        - in: body
          name: productImages
          description: productImages
          required: true
          schema:
            type: array
            items:
              $ref: '#/definitions/ProductImage'
      responses:
        '201':
          description: Created
          schema:
            type: array
            items:
              $ref: '#/definitions/ProductImage'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/api/v1/products/{product-id}/product-images/{product-image-id}':
    put:
      tags:
        - product-controller
      summary: Update a product by product image id for a product with id
      operationId: updateProductImageByIdForProductIdUsingPUT
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: product-id
          in: path
          description: product-id
          required: true
          type: integer
          format: int32
        - name: product-image-id
          in: path
          description: product-image-id
          required: true
          type: integer
          format: int32
        - in: body
          name: productImage
          description: productImage
          required: true
          schema:
            $ref: '#/definitions/ProductImage'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ProductImage'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - product-controller
      summary: Delete a product image by id for a product id
      operationId: deleteProductImageByIdUsingDELETE
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: product-id
          in: path
          description: product-id
          required: true
          type: integer
          format: int32
        - name: product-image-id
          in: path
          description: product-image-id
          required: true
          type: integer
          format: int32
      responses:
        '204':
          description: No Content
          schema:
            $ref: '#/definitions/User'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  '/api/v1/products/{product-id}/product-inventory':
    get:
      tags:
        - product-controller
      summary: Get a product inventory for a product with id
      operationId: getProductInventoryForProductIdUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: product-id
          in: path
          description: product-id
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ProductInventory'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
        - product-controller
      summary: Create a list of Product Images for a product with id
      operationId: createProductInventoryForProductIdUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: product-id
          in: path
          description: product-id
          required: true
          type: integer
          format: int32
        - in: body
          name: productInventory
          description: productInventory
          required: true
          schema:
            $ref: '#/definitions/ProductInventory'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/ProductInventory'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/api/v1/products/{product-id}/product-inventory{product-inventory-id}':
    put:
      tags:
        - product-controller
      summary: Update a product inventory product for a product with id
      operationId: updateProductInventoryForProductIdUsingPUT
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: product-id
          in: path
          description: product-id
          required: true
          type: integer
          format: int32
        - name: product-inventory-id
          in: path
          description: product-inventory-id
          required: true
          type: integer
          format: int32
        - in: body
          name: productInventory
          description: productInventory
          required: true
          schema:
            $ref: '#/definitions/ProductInventory'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ProductInventory'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - product-controller
      summary: Delete a product inventory for a product id
      operationId: deleteProductInventoryByIdNProductIdUsingDELETE
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: product-id
          in: path
          description: product-id
          required: true
          type: integer
          format: int32
        - name: product-inventory-id
          in: path
          description: product-inventory-id
          required: true
          type: integer
          format: int32
      responses:
        '204':
          description: No Content
          schema:
            $ref: '#/definitions/User'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
definitions:
...
  Offer:
    type: object
    properties:
      createdBy:
        type: integer
        format: int32
      createdDate:
        type: string
        format: date-time
      deleted:
        type: boolean
      id:
        type: integer
        format: int32
      lastModifiedBy:
        type: integer
        format: int32
      lastModifiedDate:
        type: string
        format: date-time
      name:
        type: string
      percentage:
        type: number
        format: double
      validFrom:
        type: string
        format: date-time
      validTo:
        type: string
        format: date-time
  Product:
    type: object
    properties:
      active:
        type: boolean
      caption:
        type: string
      categoryId:
        type: integer
        format: int32
      createdBy:
        type: integer
        format: int32
      createdDate:
        type: string
        format: date-time
      deleted:
        type: boolean
      description:
        type: string
      id:
        type: integer
        format: int32
      lastModifiedBy:
        type: integer
        format: int32
      lastModifiedDate:
        type: string
        format: date-time
      name:
        type: string
      price:
        type: number
        format: double
      unitOfMeasure:
        $ref: '#/definitions/UnitOfMeasure'
      viewCount:
        type: integer
        format: int32
...
  ProductImage:
    type: object
    properties:
      createdBy:
        type: integer
        format: int32
      createdDate:
        type: string
        format: date-time
      deleted:
        type: boolean
      id:
        type: integer
        format: int32
      imageUrl:
        type: string
      lastModifiedBy:
        type: integer
        format: int32
      lastModifiedDate:
        type: string
        format: date-time
      orderNo:
        type: integer
        format: int32
      productId:
        type: integer
        format: int32
  ProductInventory:
    type: object
    properties:
      availableCount:
        type: integer
        format: int32
      createdBy:
        type: integer
        format: int32
      createdDate:
        type: string
        format: date-time
      deleted:
        type: boolean
      id:
        type: integer
        format: int32
      lastModifiedBy:
        type: integer
        format: int32
      lastModifiedDate:
        type: string
        format: date-time
      productId:
        type: integer
        format: int32
      soldCount:
        type: integer
        format: int32
      status:
        type: string
        enum:
          - ACTIVE
          - OUT_OF_STOCK
          - ARRAIVING_SOON
          - PRE_ORDER
  ProductSearchRequest:
    type: object
    properties:
      categoryId:
        type: integer
        format: int32
      higherPrice:
        type: number
        format: double
      inventoryStatus:
        type: string
        enum:
          - AVAILABLE
          - OUT_OF_STOCK
      isActive:
        type: boolean
      lowerPrice:
        type: number
        format: double
      name:
        type: string
      pageNumber:
        type: integer
        format: int32
      pageSize:
        type: integer
        format: int32
      sorts:
        type: array
        items:
          $ref: '#/definitions/SortField'
      viewCount:
        type: integer
        format: int32
  Role:
    type: object
    properties:
      createdBy:
        type: integer
        format: int32
      createdDate:
        type: string
        format: date-time
      deleted:
        type: boolean
      id:
        type: integer
        format: int32
      lastModifiedBy:
        type: integer
        format: int32
      lastModifiedDate:
        type: string
        format: date-time
      name:
        type: string
...
  SortField:
    type: object
    properties:
      name:
        type: string
      sortOrder:
        type: string
        enum:
          - ASC
          - DESC
...
  UnitOfMeasure:
    type: object
    properties:
      createdBy:
        type: integer
        format: int32
      createdDate:
        type: string
        format: date-time
      deleted:
        type: boolean
      id:
        type: integer
        format: int32
      lastModifiedBy:
        type: integer
        format: int32
      lastModifiedDate:
        type: string
        format: date-time
      name:
        type: string
  User:
    type: object
    properties:
      createdBy:
        type: integer
        format: int32
      createdDate:
        type: string
        format: date-time
      deleted:
        type: boolean
      email:
        type: string
      id:
        type: integer
        format: int32
      lastModifiedBy:
        type: integer
        format: int32
      lastModifiedDate:
        type: string
        format: date-time
      mobileNo:
        type: string
      password:
        type: string
      roles:
        type: array
        items:
          $ref: '#/definitions/Role'
      status:
        type: string
        enum:
          - ACTIVE
          - INACTIVE
      userDetail:
        $ref: '#/definitions/UserDetail'
...
  UserDetail:
    type: object
    properties:
      createdBy:
        type: integer
        format: int32
      createdDate:
        type: string
        format: date-time
      deleted:
        type: boolean
      fname:
        type: string
      gender:
        type: string
        enum:
          - MALE
          - FEMALE
          - OTHER
      id:
        type: integer
        format: int32
      lastModifiedBy:
        type: integer
        format: int32
      lastModifiedDate:
        type: string
        format: date-time
      lname:
        type: string
      userId:
        type: integer
        format: int32
...