# src/main/java/org/mihy/gowma/model/ProductInventory.java
/*
 * Copyright 2017 mihy,org.
 * All rights reserved.
 */
package org.mihy.gowma.model;

public class ProductInventory extends BaseModel {

    private Integer availableCount;
    private Integer soldCount;
    private Integer productId;
    private InventoryStatus status;

    public Integer getAvailableCount() {
        return availableCount;
    }

    public void setAvailableCount(Integer availableCount) {
        this.availableCount = availableCount;
    }

    public Integer getSoldCount() {
        return soldCount;
    }

    public void setSoldCount(Integer soldCount) {
        this.soldCount = soldCount;
    }

    public InventoryStatus getStatus() {
        return status;
    }

    public void setStatus(InventoryStatus status) {
        this.status = status;
    }

    public Integer getProductId() {
        return productId;
    }

    public void setProductId(Integer productId) {
        this.productId = productId;
    }

    public enum InventoryStatus {

        ACTIVE,
        OUT_OF_STOCK,
        ARRAIVING_SOON,
        PRE_ORDER
    }
}